rationalise two entryponts
